Core module refactored to Neuron module.
As all the process, logic inside core was just to load and execute neuron

Completed with Tests passing
