+++ cerebro/nlp/__init__.py
@@ -1,4 +1,4 @@
-from .config import *
+from .constants import *
 from .parser import DataSetParser
 from .classifier import CerebroClassifier
 from .nlp import NLP
