+++ cerebro/tests/neuron/test_config.py
@@ -1,13 +1,12 @@
 import unittest
 
-import cerebro.neuron.config as cfg
-import cerebro.neuron.constants as const
+import cerebro.neuron as neu
 
 
 class TestConfigService(unittest.TestCase):
     def test_config_creation(self):
-        assert cfg.get_neurons_path() is not None
+        assert neu.cfg.get_neurons_path() is not None
 
     def test_add_path(self):
-        cfg.add_neurons_location(const.STR_DEFAULT_NEURONS_PATH)
-        assert len(cfg.get_neurons_path()) == 1
+        neu.cfg.add_neurons_location(neu.STR_DEFAULT_NEURONS_PATH)
+        assert len(neu.cfg.get_neurons_path()) == 1
