+++ cerebro/tests/neuron/test_finder.py
@@ -1,10 +1,10 @@
 import unittest
 
-from cerebro.neuron import constants as const, config as cfg, finder as fn
+import cerebro.neuron  as neu
 
 
 class TestNeuronsFinderService(unittest.TestCase):
 
     def test_find_neurons(self):
-        cfg.add_neurons_location(const.STR_DEFAULT_NEURONS_PATH)
-        assert fn.find(cfg.get_neurons_path()) is not None
+        neu.cfg.add_neurons_location(neu.STR_DEFAULT_NEURONS_PATH)
+        assert neu.find(neu.cfg.get_neurons_path()) is not None
